rmad: Get fields from cros-config-fs instead of parsing config.yaml
Since config.yaml contains large values in various complex formats,
the current json parser will have issues when parsing it with
incompatible type (base::value). Rewrite it with parsing the fs and
handling large number in our own logics to prevent the issue.

BUG=b:279528755
TEST=FEATURES=test emerge-nissa rmad

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4498190
Tested-by: Gene Chang <genechang@google.com>
Commit-Queue: Gene Chang <genechang@google.com>

+++ rmad/state_handler/update_device_info_state_handler.cc
@@ -4,6 +4,7 @@
 
 #include "rmad/state_handler/update_device_info_state_handler.h"
 
+#include <cstdint>
 #include <memory>
 #include <string>
 #include <utility>
@@ -90,13 +91,13 @@ RmadErrorCode UpdateDeviceInfoStateHandler::InitializeState() {
 
   std::string serial_number;
   std::string region;
-  int sku_id;
+  uint64_t sku_id;
   bool is_custom_label_exist;
   std::string custom_label_tag;
   std::string dram_part_number;
 
   std::vector<std::string> region_list;
-  std::vector<int> sku_id_list;
+  std::vector<uint64_t> sku_id_list;
   std::vector<std::string> custom_label_tag_list;
 
   // We reserve -1 for unmatched indexes.
@@ -191,9 +192,7 @@ RmadErrorCode UpdateDeviceInfoStateHandler::InitializeState() {
   }
 
   for (auto sku_option : sku_id_list) {
-    // We get a vector<int> from cros_config, but we set a uint64_t to cbi.
-    // Therefore, we should cast it to uint64_t here.
-    update_dev_info->add_sku_list(static_cast<uint64_t>(sku_option));
+    update_dev_info->add_sku_list(sku_option);
   }
 
   for (auto custom_label_option : custom_label_tag_list) {
