+++ rmad/utils/cros_config_utils.h
@@ -38,10 +38,10 @@ class CrosConfigUtils {
   virtual bool GetRmadConfig(RmadConfig* config) const = 0;
   virtual bool GetModelName(std::string* model_name) const = 0;
   virtual bool GetCustomLabelTag(std::string* custom_label_tag) const = 0;
-  virtual bool GetSkuId(int* sku_id) const = 0;
+  virtual bool GetSkuId(uint64_t* sku_id) const = 0;
   virtual bool GetCustomLabelTagList(
       std::vector<std::string>* custom_label_tag_list) const = 0;
-  virtual bool GetSkuIdList(std::vector<int>* sku_id_list) const = 0;
+  virtual bool GetSkuIdList(std::vector<uint64_t>* sku_id_list) const = 0;
 
   bool IsCustomLabel() const;
 };
